SQONE-462 Collapse single-post "content" component into the single-post "route" component.
